+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-# UNRELEASED
+# v1.41.0
+
+## Enhancements
+
 * Allow managing workspace and organization data retention policies by @mwudka [#801](https://github.com/hashicorp/go-tfe/pull/817)
 
 # v1.40.0